+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v1.8.0 (minor)
+
+Changes since v1.7.0:
+
+- test(sync): join the temp paths instead of combining them [patch] ([@matt-edmondson](https://github.com/matt-edmondson))
+- test(sync): drive the sync-config verbs and sync --config [patch] ([@matt-edmondson](https://github.com/matt-edmondson))
+- test(sync): cover saved configurations and the flag override [patch] ([@matt-edmondson](https://github.com/matt-edmondson))
+- sync: save named configurations of the inputs flags repeat [minor] ([@matt-edmondson](https://github.com/matt-edmondson))
+
 ## v1.7.0 (minor)
 
 Changes since v1.6.0: